+++ src/presenter/film-presenter.js
@@ -4,7 +4,6 @@ import SmartView from '../view/smart-view.js';
 import AbstractView from '../view/abstract-view.js';
 import FilmCardView from '../view/film-card-view';
 import PopupView from '../view/popup-view';
-import dayjs from 'dayjs';
 
 class FilmPresenter {
   #filmsList = null;
